all done i guess

+++ src/main/java/itmo/web/lab2/TestCase.java
@@ -22,7 +22,7 @@ public class TestCase {
         return y;
     }
 
-    public float getR() {
+    public int getR() {
         return r;
     }
